(Breaking) change: VerifyPacket behaviour
VerifyPacket previously wanted the reply packet and (optionally)
the raw packet. If only the ReplyPacket and not the actual raw
reply, then we've (tried) to reconstruct the original packet.

But we are not able to reconstruct the original packet if RADIUS
Attributes were not next to each other.

The original idea was to use the function like this:

reply = packet.CreateReply(raw_reply)
packet.VerifyReply(reply)

This breaks the case and is (unnecessary) complicated. A client
usually aways wants to verify the packet. So we can move that
into the parsing function.

With this patch VerifyPacket is supposed to be used like this:

reply = packet.VerifyReply(raw_reply)

This makes misusage of the API a lot harder. e.g. previously
something like this would be possible:

packet.VerifyReply(some_reply, another_replys_raw_data)

With the current API this problem is at least reduced, since
the user only needs to make sure to get the packet <-> raw_reply
relation correct, instead of 3 variables.

Note: we cannot do any verification in CreateReply, because that's
what should be/is used to create Replys on the server side.