(breaking) change: don't accept "secret" parameter on verify_message_authenticator
There is no good usecase to accept a secret parameter on the
verify_message_authenticator method. The key for the HMAC MUST
be the radius secret - which for any sane usage of Packet is set.

The previous implementation would fall back to use the secret
anyway. (where using the secret would be the only correct way
anyway.)